refactored content stack view
